requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
pydantic
pytelegrambotapi
